+++ core/world.ts
@@ -1,9 +1,13 @@
+import {Animate} from "./components/com_animate.js";
+import {AudioSource} from "./components/com_audio_source.js";
 import {Camera} from "./components/com_camera.js";
 import {Light} from "./components/com_light.js";
 import {Render} from "./components/com_render.js";
 import {Transform} from "./components/com_transform.js";
 
 const enum Component {
+    Animate,
+    AudioSource,
     Camera,
     Light,
     Render,
@@ -11,6 +15,8 @@ const enum Component {
 }
 
 export const enum Has {
+    Animate = 1 << Component.Animate,
+    AudioSource = 1 << Component.AudioSource,
     Camera = 1 << Component.Camera,
     Light = 1 << Component.Light,
     Render = 1 << Component.Render,
@@ -22,6 +28,8 @@ export interface World {
     Signature: Array<number>;
 
     // Component data
+    Animate: Array<Animate>;
+    AudioSource: Array<AudioSource>;
     Camera: Array<Camera>;
     Light: Array<Light>;
     Render: Array<Render>;
